PCB’s and Process Scheduling Simulation
this assignment you are going to create an interactive operating system simulation 
which supports a few basic O/S functions. The simulation will be keyboard driven, and 
will make reports to the screen. 
First of all, what do I mean by a simulation? Well, there will be no real processes doing 
any real work. There will, however, be data structures which represent processes and the 
state of the system. For example, when the simulation user issues the command to create 
a process, a PCB for the new process will be created, initialized and placed in the 
appropriate O/S queue. When the user issues the command to signal the expiry of a time 
quantum, the internal state of the simulation will react accordingly and the effects of the 
expiry will be reported to the screen. The commands issued by the simulation user will be 
on behalf of the system or the running process.